feat: add upload process bar
